requirements.txt
flask
flask-socketio
python-dotenv
google-genai
